refactor: move video label/stream leaf types to @fiftyone/utilities
